set up webpack closes #21
