+++ package.json
@@ -13,9 +13,12 @@
     "test:api": "NODE_ENV=test nodemon src/test/api.test.js | tap-spec",
     "test:coverage": "NODE_ENV=test nyc --reporter=lcov --reporter=text-lcov tape src/test/*.test.js | tap-spec && open -a \"Google Chrome\" ./coverage/lcov-report/index.html",
     "build": "node src/model/database/db_build.js",
-    "lint": "./node_modules/.bin/eslint **/*.js --ignore-pattern node_modules/"
+    "lint": "./node_modules/.bin/eslint **/*.js --ignore-pattern node_modules/",
+    "build:wp": "webpack",
+    "dev:wp": "webpack --watch"
   },
   "pre-commit": [
+    "build:wp",
     "test",
     "lint"
   ],
@@ -30,15 +33,25 @@
   },
   "homepage": "https://github.com/fac-13/stackMatch#readme",
   "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.7.0",
+    "css-loader": "^0.28.11",
     "eslint": "^4.19.1",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.11.0",
+    "mini-css-extract-plugin": "^0.4.0",
+    "node-sass": "^4.9.0",
     "nodemon": "^1.17.4",
     "nyc": "^11.7.1",
     "pre-commit": "^1.2.2",
+    "sass-loader": "^7.0.1",
+    "style-loader": "^0.21.0",
     "supertest": "^3.0.0",
     "tap-spec": "^4.1.1",
-    "tape": "^4.9.0"
+    "tape": "^4.9.0",
+    "webpack": "^4.8.3",
+    "webpack-cli": "^2.1.3"
   },
   "dependencies": {
     "axios": "^0.18.0",
