+++ public/css/style.css
@@ -105,3 +105,4 @@ a {
   background-color: #F2F2F2;
   color: #223241;
   padding: 1rem 2rem; }
+
