+++ public/dom/index.js
@@ -1,19 +1,1 @@
-/* eslint no-var: 0 */
-/* eslint prefer-arrow-callback: 0 */
-/* eslint func-names: 0 */
-/* eslint wrap-iife: 0 */
-
-function clientRequest(method, url, body, callback) {
-  var xhr = new XMLHttpRequest();
-  if (method.toLowerCase() !== 'delete') {
-    xhr.addEventListener('load', function () {
-      if (xhr.readyState === 4 && xhr.status === 200) {
-        callback(JSON.parse(xhr.responseText));
-      } else {
-        console.log('XHR error', xhr.status);
-      }
-    });
-  }
-  xhr.open(method, url, true);
-  xhr.send(body);
-}
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{configurable:!1,enumerable:!0,get:r})},n.r=function(e){Object.defineProperty(e,"__esModule",{value:!0})},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=4)}([,,,function(e,t,n){},function(e,t,n){"use strict";n(3)}]);
