+++ public/dom/profile.js
@@ -1,65 +1,1 @@
-/* eslint no-var: 0 */
-/* eslint prefer-arrow-callback: 0 */
-/* eslint func-names: 0 */
-/* eslint wrap-iife: 0 */
-
-var openModalBtn = document.querySelectorAll('.btn-open-modal');
-var closeModalBtn = document.querySelectorAll('.btn-close-modal');
-var modalIsOpen = document.querySelectorAll('.modal');
-var modalDetails = document.getElementById('modal-details');
-var modalJob = document.getElementById('modal-job');
-var modalDelete = document.getElementById('modal-delete');
-var deleteConfirmInput = document.querySelector('#delete_account_input');
-var deleteBtn = document.querySelector('#delete-account-btn');
-
-
-function closeModal() {
-  modalIsOpen.forEach((modal) => {
-    modal.style.display = 'none';
-  });
-}
-
-function openModal(profileSection) {
-  // used switch case to allow for addition profile sections, ie projects section
-  switch (profileSection) {
-    case 'details':
-      modalDetails.style.display = 'block';
-      break;
-    case 'job':
-      modalJob.style.display = 'block';
-      break;
-    case 'delete':
-      modalDelete.style.display = 'block';
-      break;
-    default:
-      break;
-  }
-  window.addEventListener('keydown', (e) => {
-    if (e.key === 'Escape') closeModal();
-  });
-}
-
-openModalBtn.forEach(btn => btn.addEventListener('click', (e) => {
-  openModal(e.target.value);
-}));
-closeModalBtn.forEach(btn => btn.addEventListener('click', (e) => {
-  closeModal();
-}));
-
-// validation
-
-// delete profile validation
-deleteConfirmInput.addEventListener('keyup', (e) => {
-  var userGithubHandle = e.target.dataset.githubhandle;
-  var confirmGithubHandle = e.target.value;
-  deleteBtn.disabled = confirmGithubHandle !== userGithubHandle;
-});
-
-// delete button delete request
-deleteBtn.addEventListener('click', (e) => {
-  e.preventDefault();
-  const xhr = new XMLHttpRequest();
-  xhr.open('DELETE', '/deleteAccount');
-  xhr.send();
-  window.location.assign('/goodbye');
-});
+!function(e){var t={};function n(o){if(t[o])return t[o].exports;var c=t[o]={i:o,l:!1,exports:{}};return e[o].call(c.exports,c,c.exports,n),c.l=!0,c.exports}n.m=e,n.c=t,n.d=function(e,t,o){n.o(e,t)||Object.defineProperty(e,t,{configurable:!1,enumerable:!0,get:o})},n.r=function(e){Object.defineProperty(e,"__esModule",{value:!0})},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=0)}([function(e,t,n){"use strict";var o=document.querySelectorAll(".btn-open-modal"),c=document.querySelectorAll(".btn-close-modal"),r=document.querySelectorAll(".modal"),l=document.getElementById("modal-details"),u=document.getElementById("modal-job"),a=document.getElementById("modal-delete"),d=document.querySelector("#delete_account_input"),i=document.querySelector("#delete-account-btn");function s(){r.forEach(function(e){e.style.display="none"})}o.forEach(function(e){return e.addEventListener("click",function(e){!function(e){switch(e){case"details":l.style.display="block";break;case"job":u.style.display="block";break;case"delete":a.style.display="block"}window.addEventListener("keydown",function(e){"Escape"===e.key&&s()})}(e.target.value)})}),c.forEach(function(e){return e.addEventListener("click",function(e){s()})}),d.addEventListener("keyup",function(e){var t=e.target.dataset.githubhandle,n=e.target.value;i.disabled=n!==t}),i.addEventListener("click",function(e){e.preventDefault();var t=new XMLHttpRequest;t.open("DELETE","/deleteAccount"),t.send(),window.location.assign("/goodbye")})}]);
